refactor: products 폴더 auctions로 변경
