+++ src/app/auctions/[auctionId]/_hooks/queries/useGetAuctionDetail.ts
