+++ src/app/auctions/new/_api/registerAuction.ts
