+++ src/app/auctions/new/_component/ProductInfo/ImageUploader/ImageUploader.tsx
@@ -2,7 +2,7 @@ import { UseFieldArrayAppend, useFormContext } from "react-hook-form";
 
 import Icon from "@/app/_component/common/Icon";
 import Toast from "@/app/_component/common/Toast";
-import { RegisterAuction } from "@/app/products/new/page";
+import { RegisterAuction } from "@/app/auctions/new/page";
 
 interface ImageUploaderProps {
   onAppend: UseFieldArrayAppend<RegisterAuction, "images">;
