+++ src/app/auctions/new/_utils/constants.ts
