+++ src/app/auctions/new/_utils/validation.ts
